build a static library for reuse of bsort
