+++ CMakeLists.txt
@@ -16,4 +16,10 @@ set(LIBRARY_OUTPUT_PATH ${CMAKE_BINARY_DIR})
 # The following folder will be included
 include_directories("${PROJECT_SOURCE_DIR}")
 
-add_executable(bsort ${CMAKE_SOURCE_DIR}/src/main.cpp)
+# build a static library
+add_library(bsort STATIC ${CMAKE_SOURCE_DIR}/src/bsort.cpp)
+
+add_executable(bsort-main ${CMAKE_SOURCE_DIR}/src/main.cpp )
+target_link_libraries(bsort-main "${LIBRARY_OUTPUT_PATH}/libbsort.a")
+set_target_properties(bsort-main PROPERTIES OUTPUT_NAME "bsort")
+add_dependencies(bsort-main bsort)
